Move styles to sub folder
