+++ src/docx/file.ts
@@ -1,10 +1,10 @@
 import { IPropertiesOptions, Properties } from "../properties";
-import { Styles } from "../styles";
-import { DefaultStylesFactory } from "../styles/factory";
 import { Document } from "./document";
 import { Media } from "./media";
 import { Numbering } from "./numbering";
 import { Paragraph } from "./paragraph";
+import { Styles } from "./styles";
+import { DefaultStylesFactory } from "./styles/factory";
 import { Table } from "./table";
 
 export class File {
