+++ src/docx/index.ts
@@ -5,3 +5,4 @@ export * from "./file";
 export * from "./numbering";
 export * from "./media";
 export * from "./drawing";
+export * from "./styles";
