+++ src/docx/styles/defaults/paragraph-properties.ts
@@ -1,5 +1,5 @@
-import { ParagraphProperties } from "../../docx/paragraph/properties";
-import { XmlComponent } from "../../docx/xml-components";
+import { ParagraphProperties } from "../../paragraph/properties";
+import { XmlComponent } from "../../xml-components";
 
 export class ParagraphPropertiesDefaults extends XmlComponent {
