+++ src/docx/styles/index.ts
@@ -1,5 +1,5 @@
-import { DocumentAttributes } from "../docx/document/document-attributes";
-import { XmlComponent } from "../docx/xml-components";
+import { DocumentAttributes } from "../document/document-attributes";
+import { XmlComponent } from "../xml-components";
 import { DocumentDefaults } from "./defaults";
 import { ParagraphStyle } from "./style";
