+++ src/index.ts
@@ -1,3 +1,2 @@
 export * from "./docx";
 export * from "./export";
-export * from "./styles";
